Make FormControl check types and remove unused props
